+++ lib/tmuxinator/project.rb
@@ -41,13 +41,11 @@ module Tmuxinator
 
     def self.load(path, options = {})
       yaml = begin
-        raw_content = File.read(path)
-
         args = options[:args] || []
         @settings = parse_settings(args)
         @args = args
 
-        content = Erubis::Eruby.new(raw_content).result(binding)
+        content = render_template(path, binding)
         YAML.safe_load(content, aliases: true)
       rescue SyntaxError, StandardError => error
         raise "Failed to parse config file: #{error.message}"
@@ -103,7 +101,7 @@ module Tmuxinator
 
     def self.render_template(template, bndg)
       content = File.read(template)
-      Erubis::Eruby.new(content).result(bndg)
+      bndg.eval(Erubi::Engine.new(content).src)
     end
 
     def windows